BSP to Unity3d.ru

+++ SimpleRPG/Assets/Scripts/Room.cs
@@ -6,9 +6,9 @@ public class Room {
 
 	public int thickness=4;
 
-	public float x1,y1,x2,y2;
+	public int x1,y1,x2,y2;
 
-	public Room(float a1,float b1,float a2,float b2){
+	public Room(int a1,int b1,int a2,int b2){
 		x1 = a1;
 		y1 = b1;
 		x2 = a2;
